+++ tests/dummy/app/controllers/components-examples/flexberry-simpleolv/settings-example.js
@@ -120,6 +120,15 @@ export default ListFormController.extend({
   refreshButton: false,
 
   /**
+    Flag: indicates whether 'flexberry-objectlistview' component is in 'defaultSortingButton' mode or not
+
+    @property defaultSortingButton
+    @type Boolean
+    @default true
+  */
+  defaultSortingButton: true,
+
+  /**
     Flag: indicates whether 'flexberry-simpleolv' component is in 'showCheckBoxInRow' mode or not.
 
     @property showCheckBoxInRow
@@ -237,6 +246,7 @@ export default ListFormController.extend({
     '  applyFilters=(action "applyFilters")<br>' +
     '  resetFilters=(action "resetFilters")<br>' +
     '  refreshButton=refreshButton<br>' +
+    '  defaultSortingButton=defaultSortingButton<br>' +
     '  filterButton=filterButton<br>' +
     '  showCheckBoxInRow=showCheckBoxInRow<br>' +
     '  showDeleteButtonInRow=showDeleteButtonInRow<br>' +
@@ -392,6 +402,12 @@ export default ListFormController.extend({
       settingType: 'boolean',
       settingDefaultValue: false,
       bindedControllerPropertieName: 'refreshButton'
+    });
+    componentSettingsMetadata.pushObject({
+      settingName: 'defaultSortingButton',
+      settingType: 'boolean',
+      settingDefaultValue: true,
+      bindedControllerPropertieName: 'defaultSortingButton'
     });
     componentSettingsMetadata.pushObject({
       settingName: 'showCheckBoxInRow',