+++ CHANGELOG.md
@@ -4,6 +4,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+* Blueprints:
+    * Add index property generation for projection attributes.
+* List components:
+    * Add option to hide default sorting button.
+* `flexberry-groupedit` component:
+    * Add check all at page button.
 * `flexberry-lookup` component:
     * Add property `autofillByLimit`.